Updating to Pixy2JavaAPI v1.2
Updating to the latest src for Pixy2JavaAPI library.  Also moving to original namespace and file locations to prepare for eventually pulling published libraries from the maven repository.  We still need Pixy2Line.getVectors method and to help debug, so sticking with a copy of the code for now.
